+++ src/Controller/MeasurementController.php
@@ -12,9 +12,9 @@ use Symfony\Component\Routing\Annotation\Route;
 
 class MeasurementController extends AbstractController
 {
-    #[Route('/', name: 'app_measurement_index', methods: ['GET'])]
     public function index(MeasurementRepository $measurementRepository): Response
     {
+        $this->denyAccessUnlessGranted('ROLE_MEASUREMENT_INDEX');
         return $this->render('measurement/index.html.twig', [
             'measurements' => $measurementRepository->findAll(),
         ]);
@@ -22,6 +22,7 @@ class MeasurementController extends AbstractController
 
     public function new(Request $request, MeasurementRepository $measurementRepository): Response
     {
+        $this->denyAccessUnlessGranted('ROLE_MEASUREMENT_CREATE');
         $measurement = new Measurement();
         $form = $this->createForm(MeasurementType::class, $measurement);
         $form->handleRequest($request);
@@ -40,6 +41,7 @@ class MeasurementController extends AbstractController
 
     public function show(Measurement $measurement): Response
     {
+        $this->denyAccessUnlessGranted('ROLE_MEASUREMENT_SHOW');
         return $this->render('measurement/show.html.twig', [
             'measurement' => $measurement,
         ]);
@@ -47,6 +49,7 @@ class MeasurementController extends AbstractController
 
     public function edit(Request $request, Measurement $measurement, MeasurementRepository $measurementRepository): Response
     {
+        $this->denyAccessUnlessGranted('ROLE_MEASUREMENT_EDIT');
         $form = $this->createForm(MeasurementType::class, $measurement);
         $form->handleRequest($request);
 
@@ -64,6 +67,7 @@ class MeasurementController extends AbstractController
 
     public function delete(Request $request, Measurement $measurement, MeasurementRepository $measurementRepository): Response
     {
+        $this->denyAccessUnlessGranted('ROLE_MEASUREMENT_DELETE');
         if ($this->isCsrfTokenValid('delete'.$measurement->getId(), $request->request->get('_token'))) {
             $measurementRepository->remove($measurement, true);
         }
